Update of opC.py to include linear interpolation
The functions for the Kramers-Krönig transform were updated to automatically use linear interpolation to approximate the k spectrum between the end of the measured spectrum and 0 cm-1. The op_constants function was also updated so that it can be run without providing an extension of the k spectrum. Now if no extension to the k spectrum is given the program will automatically use linear interpolation.